editing/input/composition-highlights.html fails on open source Mojave bots https://bugs.webkit.org/show_bug.cgi?id=207107
Reviewed by Tim Horton.

This test currently fails consistently on non-internal macOS Mojave bots; for some reason, the width of the
colon character glyph in Times font is 8.4px instead of 4.4px when using CTRunGetAdvancesPtr, causing the
composition highlights to be shifted right 4px and making the ref comparison fail.

Debugging shows that only in macOS Mojave, the CoreText framework adjusts the width of the colon glyph by 4px
in TStorageRange::SetAdvance. This has since been fixed in Catalina, so just work around this for the time being
by changing the editable text from "Test: " to just "Test ".

* editing/input/composition-highlights-expected.html:
* editing/input/composition-highlights.html:


Identifier: 220108@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@255554 268f45cc-cd09-0410-ab3c-d52691b4dbfc