[bmalloc] VMHeap can be merge into Heap https://bugs.webkit.org/show_bug.cgi?id=207410
Reviewed by Yusuke Suzuki.

VMHeap has only one member function in it and Heap is the only client of that.
No member variable is defined. It does nothing special with its context as a class.
It is safe to merge the function into Heap.

* CMakeLists.txt:
* bmalloc.xcodeproj/project.pbxproj:
* bmalloc/Heap.cpp:
(bmalloc::Heap::allocateLarge):
(bmalloc::Heap::tryAllocateLargeChunk): Moved from VMHeap.
* bmalloc/Heap.h:
* bmalloc/VMHeap.cpp: Removed.
* bmalloc/VMHeap.h: Removed.


Identifier: 220424@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@256088 268f45cc-cd09-0410-ab3c-d52691b4dbfc