[macOS] Don't fire timers when there is a pending rendering update https://bugs.webkit.org/show_bug.cgi?id=207889
Reviewed by Simon Fraser.

Source/WebCore:

* WebCore.xcodeproj/project.pbxproj:
* dom/WindowEventLoop.cpp:
(WebCore::WindowEventLoop::breakToAllowRenderingUpdate):

Add the exported interface to WindowEventLoop as the future direction is to do everything via it.
For now it just calls into ThreadTimers rather than doing anything with the event loop itself.

* dom/WindowEventLoop.h:
* platform/ThreadTimers.cpp:
(WebCore::ThreadTimers::sharedTimerFiredInternal):
(WebCore::ThreadTimers::breakFireLoopForRenderingUpdate):

If we are in a firing timer set a flag so that no more timers are fired during the current runloop cycle.

* platform/ThreadTimers.h:

Source/WebKit:

* WebProcess/WebPage/mac/TiledCoreAnimationDrawingArea.mm:
(WebKit::TiledCoreAnimationDrawingArea::scheduleRenderingUpdateRunLoopObserver):

Ensure the event loop cycles to reach the runloop observer as fast as possible.


Identifier: 220883@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@256853 268f45cc-cd09-0410-ab3c-d52691b4dbfc