[LFC][Out-of-flow] FormattingContext::Geometry::staticVerticalPositionForOutOfFlowPositioned should ignore inline boxes https://bugs.webkit.org/show_bug.cgi?id=207994 <rdar://problem/59618280>
Reviewed by Antti Koivisto.

Source/WebCore:

Test: fast/layoutformattingcontext/absolute-positioned-box-with-inline-sibling.html

https://www.w3.org/TR/CSS22/visudet.html#abs-non-replaced-height

"For the purposes of this section and the next, the term "static position" (of an element) refers, roughly,
to the position an element would have had in the normal flow."

This does not seem to be well-defined for sibling inline boxes. Let's just stick with block level boxes for now.

* layout/FormattingContextGeometry.cpp:
(WebCore::Layout::FormattingContext::Geometry::staticVerticalPositionForOutOfFlowPositioned const):

LayoutTests:

* fast/layoutformattingcontext/absolute-positioned-box-with-inline-sibling-expected.html: Added.
* fast/layoutformattingcontext/absolute-positioned-box-with-inline-sibling.html: Added.


Identifier: 220969@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257056 268f45cc-cd09-0410-ab3c-d52691b4dbfc