Fix crash in Page::updateRendering when document has no domWindow https://bugs.webkit.org/show_bug.cgi?id=207975 <rdar://56999099>
Patch by Sunny He <sunny_he@apple.com> on 2020-02-20
Reviewed by Ryosuke Niwa.

When a subframe is navigated, a new document is created and
the window is transfered asynchronously. It is possible to have
a call to updateRendering in the middle of this process, resulting
in an attempt to update a document with no window. Add a check for
this case and skip such documents.

No test, since this crash relies on sensitive timing that cannot be
consistently replicated in a test case.

* page/Page.cpp:
(WebCore::Page::updateRendering):

Identifier: 221010@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257123 268f45cc-cd09-0410-ab3c-d52691b4dbfc