[First paint] Let optional style recalcs go through while in visually-non-empty state. https://bugs.webkit.org/show_bug.cgi?id=208020 <rdar://problem/59636549>
Reviewed by Simon Fraser.

This is the final step to ensure we don't end up delaying the qualifiesAsVisuallyNonEmpty check when
the page happens to not trigger synchronous style recalcs.

Here is the optimized flow:
Optional style realc -> FrameView::styleAndRenderTreeDidChange -> qualifiesAsVisuallyNonEmpty -> Document::shouldScheduleLayout true.

This could be a slight regression in certain cases because now we let the optional style recalcs through and
only delay the redundant layouts (as opposed to delay both).

* dom/Document.cpp:
(WebCore::Document::scheduleStyleRecalc):
(WebCore::Document::shouldScheduleLayout):


Identifier: 221013@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257126 268f45cc-cd09-0410-ab3c-d52691b4dbfc