[GL][GStreamer] Instantiate GstGLContext when the shared GLContext is created https://bugs.webkit.org/show_bug.cgi?id=200626
Patch by Víctor Manuel Jáquez Leal <vjaquez@igalia.com> on 2020-02-26
Reviewed by Carlos Garcia Campos.

This patch adds GstGLContext and GstGLDisplay as part of
PlatformDisplay.

Right now one GstGLContext is created per video tag, which is a
wrapper of the GL shared context. Every time this wrapper is
created the GL context has to be current in order to initialize
the wrapper's internal state.

A better approach would be to have a single GstGLContext per GL
context, to instantiate it when the GL context is created and
current. This GstGLContext would be used for all media players.

GstGLContext and GstGLDisplay are instantiated as singletons by
PlatformDisplay. When the media player request for the
GstGLContext a GL context is instantiated, if it wasn't before,
and wraps it. PlatformDisplay adds two getters, one for
GstGLContext object, and other for GstGLDisplay. Both are used to
configure the internal GStreamer GL context in the multimedia
pipeline.

No new tests are needed.

* platform/GStreamer.cmake:
* platform/graphics/PlatformDisplay.cpp:
(WebCore::PlatformDisplay::terminateEGLDisplay):
* platform/graphics/PlatformDisplay.h:
* platform/graphics/gstreamer/GLVideoSinkGStreamer.cpp:
(requestGLContext):
(webKitGLVideoSinkChangeState):
(webKitGLVideoSinkProbePlatform):
* platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.h:
* platform/graphics/gstreamer/PlatformDisplayGStreamer.cpp: Added.
(createGstGLDisplay):
(PlatformDisplay::tryEnsureGstGLContext):
(PlatformDisplay::gstGLDisplay const):
(PlatformDisplay::gstGLContext const):

Identifier: 221181@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257463 268f45cc-cd09-0410-ab3c-d52691b4dbfc